Типизация PostView с сопутствующими фиксами типов зависимых компонентов
