+++ src/components/PostComment/types.ts
@@ -1,8 +1,8 @@
 import { ReactElement } from 'react';
-import { PostData, UserData } from '../../types/types';
+import { PostData, PostUserData } from '../../types/types';
 
 export interface PostCommentProps {
   data: PostData;
   contentArray: ReactElement[];
-  postOwner: UserData;
+  postOwner: PostUserData;
 }
